requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
numpy
pillow
imageio
tensorboard
sklearn
matplotlib
future
pyyaml
packaging
setproctitle
tqdm
scikit-image
